+++ test/objects.js
@@ -1,4 +1,7 @@
+/* eslint-disable global-require, import/no-extraneous-dependencies, import/no-unresolved, max-len */
+
 import test from 'ava';
+
 let timm;
 if (process.env.TEST_MINIFIED_LIB) {
   timm = require('../lib/timm.min');
@@ -12,56 +15,56 @@ const OBJ  = { a: 1, b: 2, d: { d1: 3, d2: 4, b: { b: { b: 4 } } }, e: { e1: 'fo
 //------------------------------------------------
 // getIn()
 //------------------------------------------------
-test('getIn: object root: shallow', t => {
+test('getIn: object root: shallow', (t) => {
   t.is(timm.getIn(OBJ, ['a']), 1);
 });
 
-test('getIn: object root: deep', t => {
+test('getIn: object root: deep', (t) => {
   t.is(timm.getIn(OBJ, ['d', 'b', 'b', 'b']), 4);
 });
 
-test('getIn: array root: shallow', t => {
+test('getIn: array root: shallow', (t) => {
   t.deepEqual(timm.getIn(ARR, [1]), { a: 2 });
 });
 
-test('getIn: array root: deep', t => {
+test('getIn: array root: deep', (t) => {
   t.is(timm.getIn(ARR, [2, 'd', 'd2']), 5);
 });
 
-test('getIn: array root: deep - null value', t => {
+test('getIn: array root: deep - null value', (t) => {
   t.is(timm.getIn(ARR, [2, 'd', 'd3']), null);
 });
 
-test('getIn: should return the object for an empty array', t => {
+test('getIn: should return the object for an empty array', (t) => {
   t.is(timm.getIn(ARR, []), ARR);
   t.is(timm.getIn(OBJ, []), OBJ);
 });
 
-test('getIn: should return undefined for an unknown path', t => {
+test('getIn: should return undefined for an unknown path', (t) => {
   t.is(timm.getIn(OBJ, ['d', 'j']), undefined);
   t.is(timm.getIn(ARR, [23]), undefined);
 });
 
-test('getIn: should return undefined for a null/undefined object', t => {
+test('getIn: should return undefined for a null/undefined object', (t) => {
   t.is(timm.getIn(null, ['a']), undefined);
   t.is(timm.getIn(undefined, ['a']), undefined);
 });
 
-test('getIn: should throw for an unspecified path', t => {
+test('getIn: should throw for an unspecified path', (t) => {
   t.throws(() => timm.getIn(OBJ));
 });
 
 //------------------------------------------------
 // set()
 //------------------------------------------------
-test('set: changing', t => {
+test('set: changing', (t) => {
   const obj2 = timm.set(OBJ, 'b', 5);
   t.is(OBJ.b, 2);
   t.not(obj2, OBJ);
   t.is(obj2.b, 5);
 });
 
-test('set: should return the same object when it hasn\'t changed', t => {
+test('set: should return the same object when it hasn\'t changed', (t) => {
   const obj2 = timm.set(OBJ, 'b', 2);
   t.is(obj2, OBJ);
 });
@@ -69,7 +72,7 @@ test('set: should return the same object when it hasn\'t changed', t => {
 //------------------------------------------------
 // setIn()
 //------------------------------------------------
-test('setIn: level 2: with change', t => {
+test('setIn: level 2: with change', (t) => {
   const obj2 = timm.setIn(OBJ, ['d', 'd1'], 4);
   t.is(OBJ.d.d1, 3);
   t.not(obj2, OBJ);
@@ -79,7 +82,7 @@ test('setIn: level 2: with change', t => {
   t.is(obj2.e, OBJ.e);
 });
 
-test('setIn: level 2: should return the same object when it hasn\'t changed', t => {
+test('setIn: level 2: should return the same object when it hasn\'t changed', (t) => {
   const obj2 = timm.setIn(OBJ, ['d', 'd1'], 3);
   t.is(OBJ.d.d1, 3);
   t.is(obj2, OBJ);
@@ -88,7 +91,7 @@ test('setIn: level 2: should return the same object when it hasn\'t changed', t 
   t.is(obj2.e, OBJ.e);
 });
 
-test('setIn: level 2: should not convert arrays to objects', t => {
+test('setIn: level 2: should not convert arrays to objects', (t) => {
   const obj2 = timm.setIn(OBJ, ['arr', 2], 'e');
   t.is(OBJ.arr.length, 2);
   t.not(obj2, OBJ);
@@ -97,7 +100,7 @@ test('setIn: level 2: should not convert arrays to objects', t => {
   t.is(obj2.arr[2], 'e');
 });
 
-test('setIn: deeper: with change', t => {
+test('setIn: deeper: with change', (t) => {
   const obj2 = timm.setIn(OBJ, ['d', 'b', 'b', 'b'], 3);
   t.is(OBJ.d.b.b.b, 4);
   t.is(obj2.d.b.b.b, 3);
@@ -108,7 +111,7 @@ test('setIn: deeper: with change', t => {
   t.is(obj2.e, OBJ.e);
 });
 
-test('setIn: deeper: should return the same object when it hasn\'t changed', t => {
+test('setIn: deeper: should return the same object when it hasn\'t changed', (t) => {
   const obj2 = timm.setIn(OBJ, ['d', 'b', 'b', 'b'], 4);
   t.is(OBJ.d.b.b.b, 4);
   t.is(obj2, OBJ);
@@ -118,17 +121,17 @@ test('setIn: deeper: should return the same object when it hasn\'t changed', t =
   t.is(obj2.e, OBJ.e);
 });
 
-test('setIn: should create nested objects for unknown paths', t => {
+test('setIn: should create nested objects for unknown paths', (t) => {
   const obj2 = timm.setIn(OBJ, ['unknown', 'long', 'path'], 3);
   t.is(obj2.unknown.long.path, 3);
 });
 
-test('setIn: should return the value if the path is empty', t => {
+test('setIn: should return the value if the path is empty', (t) => {
   const obj2 = timm.setIn(OBJ, [], { a: 3 });
   t.deepEqual(obj2, { a: 3 });
 });
 
-test('setIn: should allow unsetting an attribute', t => {
+test('setIn: should allow unsetting an attribute', (t) => {
   const obj2 = timm.setIn(OBJ, ['d', 'd1'], undefined);
   t.not(obj2, OBJ);
   t.is(obj2.d.d1, undefined);
@@ -137,7 +140,7 @@ test('setIn: should allow unsetting an attribute', t => {
 //------------------------------------------------
 // updateIn()
 //------------------------------------------------
-test('updateIn: with changes', t => {
+test('updateIn: with changes', (t) => {
   const obj2 = timm.updateIn(OBJ, ['e', 'e1'], val => `${val}x`);
   t.is(OBJ.e.e1, 'foo');
   t.not(obj2, OBJ);
@@ -145,12 +148,12 @@ test('updateIn: with changes', t => {
   t.is(obj2.d, OBJ.d);
 });
 
-test('updateIn: should return the same object when it hasn\'t changed', t => {
+test('updateIn: should return the same object when it hasn\'t changed', (t) => {
   const obj2 = timm.updateIn(OBJ, ['e', 'e1'], val => val);
   t.is(obj2, OBJ);
 });
 
-test('updateIn: should create nested objects for unknown paths', t => {
+test('updateIn: should create nested objects for unknown paths', (t) => {
   const obj2 = timm.updateIn(OBJ, ['unknown', 'long', 'path'], () => 3);
   t.is(obj2.unknown.long.path, 3);
 });
@@ -158,7 +161,7 @@ test('updateIn: should create nested objects for unknown paths', t => {
 //------------------------------------------------
 // merge()
 //------------------------------------------------
-test('merge: with changes', t => {
+test('merge: with changes', (t) => {
   const obj2 = timm.merge(OBJ, { b: 4, c: 3 });
   t.is(OBJ.b, 2);
   t.is(OBJ.c, undefined);
@@ -170,48 +173,48 @@ test('merge: with changes', t => {
   t.is(obj2.e, OBJ.e);
 });
 
-test('merge: with more than 6 args', t => {
+test('merge: with more than 6 args', (t) => {
   const obj2 = timm.merge({ a: 1 }, { b: 2 }, { c: 3 }, { d: 4 }, { e: 5 }, { f: 6 }, { g: 7 });
   t.deepEqual(obj2, { a: 1, b: 2, c: 3, d: 4, e: 5, f: 6, g: 7 });
 });
 
-test('merge: should return the same object when merged with undefined', t => {
+test('merge: should return the same object when merged with undefined', (t) => {
   const obj2 = timm.merge(OBJ, undefined);
   t.is(obj2, OBJ);
 });
 
-test('merge: should ignore undefined attributes in subsequent args', t => {
+test('merge: should ignore undefined attributes in subsequent args', (t) => {
   const obj2 = timm.merge(OBJ, { a: undefined });
   t.is(obj2, OBJ);
   t.is(OBJ.a, 1);
 });
 
-test('merge: should NOT ignore null attributes in subsequent args', t => {
+test('merge: should NOT ignore null attributes in subsequent args', (t) => {
   const obj2 = timm.merge(OBJ, { g: null });
   t.not(obj2, OBJ);
   t.is(obj2.g, null);
 });
 
-test('merge: should return the same object when merged with null', t => {
+test('merge: should return the same object when merged with null', (t) => {
   const obj2 = timm.merge(OBJ, null);
   t.is(obj2, OBJ);
 });
 
-test('merge: should return the same object when merged with an empty object', t => {
+test('merge: should return the same object when merged with an empty object', (t) => {
   const obj2 = timm.merge(OBJ, {});
   t.is(obj2, OBJ);
 });
 
-test('merge: should return the same object when it hasn\'t changed', t => {
+test('merge: should return the same object when it hasn\'t changed', (t) => {
   const obj2 = timm.merge(OBJ, { b: 2, d: OBJ.d });
   t.is(obj2, OBJ);
 });
 
-test('merge: should throw with no args', t => {
+test('merge: should throw with no args', (t) => {
   t.throws(timm.merge);
 });
 
-test('merge: multiple: with changes', t => {
+test('merge: multiple: with changes', (t) => {
   const obj2 = timm.merge(OBJ, { b: 4 }, { c: 3 }, { b: 7 });
   t.is(OBJ.b, 2);
   t.is(OBJ.c, undefined);
@@ -223,12 +226,12 @@ test('merge: multiple: with changes', t => {
   t.is(obj2.e, OBJ.e);
 });
 
-test('merge: multiple: should return the same object when merged with undefined, null and empty objects', t => {
+test('merge: multiple: should return the same object when merged with undefined, null and empty objects', (t) => {
   const obj2 = timm.merge(OBJ, undefined, null, {}, null, undefined);
   t.is(obj2, OBJ);
 });
 
-test('merge: multiple: should return the same object when it hasn\'t changed', t => {
+test('merge: multiple: should return the same object when it hasn\'t changed', (t) => {
   const obj2 = timm.merge(OBJ, { b: 2 }, { d: OBJ.d }, { c: undefined });
   t.is(obj2, OBJ);
 });
@@ -236,7 +239,7 @@ test('merge: multiple: should return the same object when it hasn\'t changed', t
 //------------------------------------------------
 // mergeIn()
 //------------------------------------------------
-test('mergeIn: with changes', t => {
+test('mergeIn: with changes', (t) => {
   const obj2 = timm.mergeIn(OBJ, ['d', 'b', 'b'], { a: 3, c: 5 });
   t.deepEqual(OBJ.d.b.b, { b: 4 });
   t.not(obj2, OBJ);
@@ -244,12 +247,12 @@ test('mergeIn: with changes', t => {
   t.is(obj2.e, OBJ.e);
 });
 
-test('mergeIn: should create nested objects for unknown paths', t => {
+test('mergeIn: should create nested objects for unknown paths', (t) => {
   const obj2 = timm.mergeIn(OBJ, ['unknown', 'path'], { d: 4 });
   t.deepEqual(obj2.unknown.path, { d: 4 });
 });
 
-test('mergeIn: with more than 7 args', t => {
+test('mergeIn: with more than 7 args', (t) => {
   const obj2 = timm.mergeIn({ a: 1 }, [], { b: 2 }, { c: 3 }, { d: 4 }, { e: 5 }, { f: 6 }, { g: 7 });
   t.deepEqual(obj2, { a: 1, b: 2, c: 3, d: 4, e: 5, f: 6, g: 7 });
 });
@@ -257,7 +260,7 @@ test('mergeIn: with more than 7 args', t => {
 //------------------------------------------------
 // omit()
 //------------------------------------------------
-test('omit: with changes (single attribute)', t => {
+test('omit: with changes (single attribute)', (t) => {
   const obj2 = timm.omit(OBJ, 'a');
   t.is(obj2.a, undefined);
   t.is(obj2.b, 2);
@@ -265,7 +268,7 @@ test('omit: with changes (single attribute)', t => {
   t.deepEqual(obj2.d, OBJ.d);
 });
 
-test('omit: with changes (multiple attributes)', t => {
+test('omit: with changes (multiple attributes)', (t) => {
   const obj2 = timm.omit(OBJ, ['a', 'b']);
   t.is(obj2.a, undefined);
   t.is(obj2.b, undefined);
@@ -273,7 +276,7 @@ test('omit: with changes (multiple attributes)', t => {
   t.deepEqual(obj2.d, OBJ.d);
 });
 
-test('omit: should return the same object when it hasn\'t changed', t => {
+test('omit: should return the same object when it hasn\'t changed', (t) => {
   const obj2 = timm.omit(OBJ, 'z');
   t.deepEqual(obj2, OBJ);
 });
@@ -281,7 +284,7 @@ test('omit: should return the same object when it hasn\'t changed', t => {
 //------------------------------------------------
 // addDefaults()
 //------------------------------------------------
-test('addDefaults: with changes', t => {
+test('addDefaults: with changes', (t) => {
   const obj2 = timm.addDefaults(OBJ, { b: 4, c: 3 });
   t.is(OBJ.b, 2);
   t.is(OBJ.c, undefined);
@@ -293,38 +296,38 @@ test('addDefaults: with changes', t => {
   t.is(obj2.e, OBJ.e);
 });
 
-test('addDefaults: with changes (null attribute)', t => {
+test('addDefaults: with changes (null attribute)', (t) => {
   const obj2 = timm.addDefaults(OBJ, { f: null });
   t.not(obj2, OBJ);
   t.is(obj2.f, null);
 });
 
-test('addDefaults: with more than 6 args', t => {
+test('addDefaults: with more than 6 args', (t) => {
   const obj2 = timm.addDefaults({ a: 1 }, { b: 2 }, { c: 3 }, { d: 4 }, { e: 5 }, { f: 6 }, { g: 7 });
   t.deepEqual(obj2, { a: 1, b: 2, c: 3, d: 4, e: 5, f: 6, g: 7 });
 });
 
-test('addDefaults: should return the same object when combined with undefined', t => {
+test('addDefaults: should return the same object when combined with undefined', (t) => {
   const obj2 = timm.addDefaults(OBJ, undefined);
   t.is(obj2, OBJ);
 });
 
-test('addDefaults: should return the same object when combined with null', t => {
+test('addDefaults: should return the same object when combined with null', (t) => {
   const obj2 = timm.addDefaults(OBJ, null);
   t.is(obj2, OBJ);
 });
 
-test('addDefaults: should return the same object when combined with an empty object', t => {
+test('addDefaults: should return the same object when combined with an empty object', (t) => {
   const obj2 = timm.addDefaults(OBJ, {});
   t.is(obj2, OBJ);
 });
 
-test('addDefaults: should return the same object when it hasn\'t changed', t => {
+test('addDefaults: should return the same object when it hasn\'t changed', (t) => {
   const obj2 = timm.addDefaults(OBJ, { b: 2, d: OBJ.d });
   t.is(obj2, OBJ);
 });
 
-test('addDefaults: multiple: with changes', t => {
+test('addDefaults: multiple: with changes', (t) => {
   const obj2 = timm.addDefaults(OBJ, { b: 4 }, { c: 3 }, { b: 7 }, { c: 6 });
   t.is(OBJ.b, 2);
   t.is(OBJ.c, undefined);
@@ -336,12 +339,12 @@ test('addDefaults: multiple: with changes', t => {
   t.is(obj2.e, OBJ.e);
 });
 
-test('addDefaults: multiple: should return the same object when combined with undefined, null and empty objects', t => {
+test('addDefaults: multiple: should return the same object when combined with undefined, null and empty objects', (t) => {
   const obj2 = timm.addDefaults(OBJ, undefined, null, {}, null, undefined);
   t.is(obj2, OBJ);
 });
 
-test('addDefaults: multiple: should return the same object when it hasn\'t changed', t => {
+test('addDefaults: multiple: should return the same object when it hasn\'t changed', (t) => {
   const obj2 = timm.addDefaults(OBJ, { b: 2 }, { d: OBJ.d }, { c: undefined });
   t.is(obj2, OBJ);
 });