+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+* Add **removeFirst(), removeLast()**
+
 ## 1.1.3 (October 19, 2016)
 
 * Make user-facing Flow type definitions inherit from internal ones (to avoid divergence).